Moved HeadGetter and HeadRequester to util/heads package
