+++ src/main/java/world/bentobox/bentobox/util/heads/HeadRequester.java
@@ -1,4 +1,4 @@
-package world.bentobox.bentobox.util;
+package world.bentobox.bentobox.util.heads;
 
 import world.bentobox.bentobox.api.panels.PanelItem;
